se incluyen las validaciones de id, titulo y descripcion para actualizacion y de id para eliminación

+++ src/main/java/com/leruizr/toDoApp/service/TaskService.java
@@ -12,7 +12,6 @@ public class TaskService {
     @Autowired
     private TaskRepository taskRepository;
 
-    // Método para crear una nueva tarea
     public Task createTask(Task task) {
 
         LocalDateTime now = LocalDateTime.now();
@@ -22,37 +21,60 @@ public class TaskService {
         return taskRepository.save(task);
     }
 
-    // Método para consultar todas las tareas
-
     public Iterable<Task> getAllTasks() {
         return taskRepository.findAll();
     }
 
-    // Método para consultar una tarea por su id
 
     public Task getTaskById(Long id) {
         return taskRepository.findById(id).orElse(null);
     }
 
-    // Método para actualizar una tarea por su id
 
     public Task updateTask(Long id, Task task) {
         Task taskToUpdate = taskRepository.findById(id).orElse(null);
 
         if (taskToUpdate == null) {
-            return null;
+            throw new IllegalArgumentException("No se actualizó porque el ID no existe");
         }
 
-        taskToUpdate.setTitle(task.getTitle());
-        taskToUpdate.setDescription(task.getDescription());
+        if (task.getTitle() == null || task.getTitle().trim().isEmpty()) {
+            throw new IllegalArgumentException("No se actualizó porque falta el título");
+        }
+
+        if (task.getDescription() == null || task.getDescription().trim().isEmpty()) {
+            throw new IllegalArgumentException("No se actualizó porque falta la descripción");
+        }
+
+        // Actualizar solo si hay un cambio real en los valores
+        boolean isUpdated = false;
+
+        if (!task.getTitle().equals(taskToUpdate.getTitle())) {
+            taskToUpdate.setTitle(task.getTitle());
+            isUpdated = true;
+        }
+
+        if (!task.getDescription().equals(taskToUpdate.getDescription())) {
+            taskToUpdate.setDescription(task.getDescription());
+            isUpdated = true;
+        }
+
+        if (!isUpdated) {
+            throw new IllegalArgumentException("No se realizaron cambios porque los valores no cambiaron");
+        }
+
         taskToUpdate.setUpdatedAt(LocalDateTime.now());
 
         return taskRepository.save(taskToUpdate);
     }
 
-    // Método para eliminar una tarea por su id
-
     public void deleteTask(Long id) {
+        Task taskToDelete = taskRepository.findById(id).orElse(null);
+
+        if (taskToDelete == null) {
+            throw new IllegalArgumentException("No se eliminó porque el ID no existe");
+        }
+
         taskRepository.deleteById(id);
     }
